// frontend/src/app/modules/landing/employer/models/job-circular.ts
export interface JobCircular {
    title: string;
    description: string;
    requirement_details: string;
    additional_requirements: string;
    responsibilities: string;
    category: any;
    job_type: any;
    experience_level: any;
    salary_upper_limit: any;
    salary_lower_limit: any;
    experience_lower_limit: any;
    experience_upper_limit: any;
    other_benefits: string;
    location: string;
    currency: string;
    company_site_link: any;
    number_of_vacancies: any;
    application_deadline: any;
}